fix(cuda): align QK256 I2_S GEMV layout (#4024)
## Summary
- align the CUDA QK256 GEMV unpacking with BitNet.cpp/GGML I2_S grouped
bitplane layout
- use the canonical I2_S code map `[-1, 0, 1, 0]` instead of the stale
contiguous `[-2, -1, 1, 2]` map
- update the CUDA QK256 test oracle/fixtures to match the CPU scalar
oracle

## Validation
- cargo test --locked -p bitnet-kernels --no-default-features --features
cuda --lib qk256 -- --nocapture
- $env:PATH="C:\Program Files\NVIDIA GPU Computing
Toolkit\CUDA\v12.9\bin;$env:PATH"; $env:BITNET_RUN_CUDA_QK256_GEMV='1';
cargo test --locked -p bitnet-kernels --no-default-features --features
cuda --lib
cuda::qk256_gemv::tests::test_cuda_qk256_gemv_live_parity_opt_in --
--nocapture
- cargo check --locked -p bitnet-cli --no-default-features --features
cpu,cuda,full-cli
- cargo run --locked --release -p bitnet-cli --no-default-features
--features cpu,cuda,full-cli -- ask --device nvidia-rtx-5070-ti-cuda
--model
D:\Code\Rust\BitNet\models\microsoft-bitnet-b1.58-2B-4T-gguf\ggml-model-i2_s.gguf
--tokenizer
D:\Code\Rust\BitNet\models\microsoft-bitnet-b1.58-2B-4T\tokenizer.json
--question "What is 2+2? Answer with only the number." --max-new-tokens
8 --temperature 0 --strict-cuda --receipt-out
target\bitnet\receipts\cuda-answer-readiness\strict-cuda-ask-math-after-qk256-fix.json
- cargo run --locked --release -p bitnet-cli --no-default-features
--features cpu,cuda,full-cli -- answer-corpus --device
nvidia-rtx-5070-ti-cuda --model
D:\Code\Rust\BitNet\models\microsoft-bitnet-b1.58-2B-4T-gguf\ggml-model-i2_s.gguf
--tokenizer
D:\Code\Rust\BitNet\models\microsoft-bitnet-b1.58-2B-4T\tokenizer.json
--corpus ci\quality\bitnet-answer-corpus.yaml
--per-prompt-timeout-seconds 300 --dump-logit-steps 1 --logits-topk 20
--fail-on-quality --json-out
target\bitnet\receipts\cuda-answer-readiness\cuda-answer-corpus-after-qk256-fix.json
- cargo fmt -p bitnet-kernels -- --check
- git diff --check

## Result
- strict CUDA ask answers `4` with generated token IDs `220,19,128009`
- CUDA answer corpus passes all five deterministic cases with
`selected_backend=nvidia-rtx-5070-ti-cuda`, `runtime_api=cuda`,
`fallback_used=false`, and `selected_kernel=qk256_gemv_cuda`
- no speedup claim

+++ crates/bitnet-ffi/src/config.rs
@@ -283,7 +283,7 @@ impl BitNetCInferenceConfig {
         config.repetition_penalty = self.repetition_penalty;
 
         if self.seed != 0 {
-            config = config.with_seed(self.seed);
+            config = config.with_seed(self.seed as u64);
         }
 
         config
